--- a/applications/webrtc_video_server/webrtc_server.py
+++ b/applications/webrtc_video_server/webrtc_server.py
@@ -29,8 +29,57 @@ from operators.webrtc_server.webrtc_server_op import WebRTCServerOp
 ROOT = os.path.dirname(__file__)
 
 
+def parse_ice_strings(ice_server_strings):
+    """
+    Convert a list of ICE server strings into a list of dictionaries of the iceServers format.
+    The iceServers format is from https://developer.mozilla.org/en-US/docs/Web/API/RTCPeerConnection/RTCPeerConnection#iceservers
+
+    Parameters:
+    - ice_server_strings (list): A list of ICE server strings. Each string is expected to be
+      in the format 'protocol:host:port[username:credential]' where '[username:credential]' is needed for TURN server.
+
+
+    Returns:
+    list: A list of dictionaries, where each dictionary represents an ICE server.
+          Each dictionary has at least a 'urls' key, and may include 'username' and 'credential' keys.
+
+    Example:
+    If input_list is ['turn:10.0.0.131:3478[admin:admin]', 'stun:stun.l.google.com:19302'],
+    the output will be [
+        {
+            'urls': 'turn:10.0.0.131:3478',
+            'username': 'admin',
+            'credential': 'admin',
+        },
+        {
+            'urls': 'stun:stun.l.google.com:19302',
+        },
+    ]
+    """
+    ice_server_list = []
+    for ice_string in ice_server_strings:
+        if "[" in ice_string:
+            parts = ice_string.split("[")
+            url = parts[0]
+            creds = parts[1].split(":")
+            if len(creds) != 2:
+                raise ValueError(
+                    "Expected ice-server with credentials to be in form of turn:<ip>:<port>[<username>:<password>] "
+                )
+            username = creds[0]
+            password = creds[1][:-1]
+            ice_server = {"urls": url, "username": username, "credential": password}
+        else:
+            ice_server = {"urls": ice_string}
+        ice_server_list.append(ice_server)
+    logging.debug(f"Parsed ice strings into {ice_server_list}")
+    return ice_server_list
+
+
 class WebAppThread(Thread):
-    def __init__(self, webrtc_server_op, host, port, cert_file=None, key_file=None):
+    def __init__(
+        self, webrtc_server_op, host, port, cert_file=None, key_file=None, ice_server=None
+    ):
         super().__init__()
         self._webrtc_server_op = webrtc_server_op
         self._host = host
@@ -42,11 +91,16 @@ class WebAppThread(Thread):
         else:
             self._ssl_context = None
 
+        self._ice_servers = []
+        if ice_server:
+            self._ice_servers = parse_ice_strings(ice_server)
+
         app = web.Application()
         app.on_shutdown.append(self._on_shutdown)
         app.router.add_get("/", self._index)
         app.router.add_get("/client.js", self._javascript)
         app.router.add_post("/offer", self._offer)
+        app.router.add_get("/iceServers", self._get_ice_servers)
 
         self._runner = web.AppRunner(app)
 
@@ -60,6 +114,10 @@ class WebAppThread(Thread):
     async def _javascript(self, request):
         content = open(os.path.join(ROOT, "client.js"), "r").read()
         return web.Response(content_type="application/javascript", text=content)
+
+    async def _get_ice_servers(self, request):
+        logging.info(f"Available ice servers are {json.dumps(self._ice_servers)}")
+        return web.Response(content_type="application/json", text=json.dumps(self._ice_servers))
 
     async def _offer(self, request):
         params = await request.json()
@@ -126,6 +184,7 @@ class WebRTCServerApp(holoscan.core.Application):
             self._cmdline_args.port,
             self._cmdline_args.cert_file,
             self._cmdline_args.key_file,
+            self._cmdline_args.ice_server,
         )
         self._web_app_thread.start()
 
@@ -139,6 +198,12 @@ if __name__ == "__main__":
         "--port", type=int, default=8080, help="Port for HTTP server (default: 8080)"
     )
     parser.add_argument("--verbose", "-v", action="store_true")
+    parser.add_argument(
+        "--ice-server",
+        action="append",
+        help="ICE server config in the form of `turn:<ip>:<port>[<username>:<password>]` or `stun:<ip>:<port>`. "
+        "This option can be specified multiple times to add multiple ICE servers.",
+    )
     cmdline_args = parser.parse_args()
 
     if cmdline_args.verbose: