@MaggieXJZhang 2023-11-27
I've only added docker instructions. Please let me know if you'd like to see K8S instructions (i.e. helm chart, deployment steps) in this PR as well or not.