@slbryson 2023-11-28
Minor comment: In the first performance table, should the fourth column be "..After Reorder.."